remove fixtures folder
